feat: use updated clair-action/claircore
We have found[^1] that the Red Hat VEX[^2] data consistently contains
the release date information, and the OVAL[^3] data consistently does
not. Recently[^4] claircore introduced the Red Hat VEX updater and the
clair-action was updated[^5] to use that version.

This updates to clair-action 0.0.9 which updates claircore to 1.5.30]
containing the needed Red Hat VEX support.

In addition to that, we needed an update to claircore 1.5.31 which
contains a solution for robustness of processing of the VEX files[^6].

Resolves: https://issues.redhat.com/browse/EC-835

[^1]: https://issues.redhat.com/browse/EC-812
[^2]: https://www.redhat.com/en/blog/red-hat-vex-files-cves-are-now-generally-available
[^3]: https://access.redhat.com/articles/221883/
[^4]: https://issues.redhat.com/browse/CLAIRDEV-28
[^5]: https://issues.redhat.com/browse/EC-854
[^6]: https://github.com/quay/claircore/pull/1410